testing/python/python_11/pyproject.toml
[project]
name = "python-11"
version = "0.1.0"
description = ""
authors = [
    {name = "haxscramper",email = "haxscramper@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "piexif (>=1.1.3,<2.0.0)",
    "click (>=8.3.0,<9.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
